fix broken test and restructure edifact module
