+++ nhais-adaptor/edifact/incoming/message.py
